requirements.txt
amqp==5.2.0                                                                               
anyio==4.6.0                                                                                        
async-timeout==4.0.3                                                                                
attrs==23.2.0                                                                                       
autobahn==24.4.2                                                                                    
Automat==24.8.1                                                                                     
billiard==4.2.0                                                                                     
celery==5.4.0                                                                                       
certifi==2024.7.4                                                                                   
cffi==1.17.1                                                                                        
channels==4.2.0                                                                                     
channels_redis==4.2.1                                                                               
charset-normalizer==3.3.2                                                                           
click==8.1.7                                                                                        
click-didyoumean==0.3.1                                                                             
click-plugins==1.1.1                                                                                
click-repl==0.3.0                                                                                   
colorama==0.4.6                                                                                     
constantly==23.10.4                                                                                 
cryptography==44.0.0                                                                                
daphne==4.1.2                                                                                       
django-guardian==2.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-recursive==0.1.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.27.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5                                                                           
httpx==0.27.2                                                                                       
hyperlink==21.0.0                                                                                   
idna==3.7                                                                                           
incremental==24.7.2                                                                                 
inflection==0.5.1                                                                                   
jalali_core==1.0.0                                                                       
jdatetime==5.0.0                                                                                    
jsonschema==4.23.0                                                                                  
jsonschema-specifications==2023.12.1                                                                
mysqlclient==2.2.4                                                                                  
outcome==1.3.0.post0
packaging==24.2
phonenumbers==8.13.40
pika==1.3.2
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
selenium==4.27.1
service-identity==24.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
tomli==2.2.1
trio==0.28.0
trio-websocket==0.11.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2 
tzdata==2024.1            
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.2